added new project in which i worked with projection and entity relationship of spring boot

+++ hospital-management/src/main/java/com/example/repository/DepartmentRepository.java
@@ -3,5 +3,7 @@ package com.example.repository;
 import com.example.entity.Department;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepartmentRepository extends JpaRepository<Department,Integer> {
+import java.util.Set;
+
+public interface DepartmentRepository extends JpaRepository<Department, Long> {
 }
